refactor: move delegation table to named_keys
If we store the delegations under the known urefs of PoS,
we can effectively query them in host and can impose the cost
to the host not to the contract.
Moreover, delegation table can potentially grow large,
if we store them under Key::Local, serialization cost could be
significant performance loss.